ingestion/requirements.txt
google-cloud-storage
google-cloud-secret-manager
google-cloud-pubsub
google-cloud-logging
requests